requirements.txt
fastapi
uvicorn[standard]
python-multipart
SQLAlchemy
pydantic
pymupdf
pillow
pytesseract
faiss-cpu
sentence-transformers
numpy
openai
python-dotenv